requirements.txt
# Core Dependencies
numpy>=1.20.0
scipy>=1.5.0
pandas>=2.0.0  # For historical data integration and data processing
scikit-learn>=1.3.0  # For StandardScaler and data preprocessing
qiskit>=0.40.0
qiskit-aer>=0.12.0  # Required for Qiskit 2.x quantum simulation

# Social Media & Sentiment Analysis
snscrape>=0.7.0.20230622  # For live X/Twitter sentiment analysis
vaderSentiment>=3.3.2  # For sentiment analysis of social media content

# Decentralized Storage
ipfshttpclient>=0.8.0a2  # For decentralized KPI dashboard publishing and trauma logging
arweave-python>=2.0.0  # For permanent storage of trauma logs on Arweave (optional)

# Zero-Knowledge Proofs
# Note: snarkjs requires Node.js and npm: npm install -g snarkjs
# Circuit files (eas.wasm, eas.zkey) must be generated separately using circom

# RSS Feed Monitoring
feedparser>=6.0.10  # For WHO RSS feed monitoring

# Progress & Utilities
tqdm>=4.65.0
pyyaml>=6.0
matplotlib>=3.7.0

# Web Framework
flask>=2.3.0
requests>=2.31.0
python-dotenv>=1.0.0

# AI/ML Models
openai>=1.0.0  # For Whisper voice consent processing
transformers>=4.30.0  # For Whisper model and LoRA support
peft>=0.3.0  # For LoRA (Low-Rank Adaptation) adapters
torch>=2.0.0  # PyTorch for model training

# Security & Encryption
cryptography>=41.0.0  # For encrypting federated learning deltas

# Audio Processing
soundfile>=0.12.0  # For audio processing (optional, for better audio handling)

# Blockchain Integration (Optional)
web3>=6.0.0  # For SurvivorDAO blockchain integration (optional)
